% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analysis_augment.R
\name{apply_rec}
\alias{apply_rec}
\title{Apply Preprocessing Recipe}
\usage{
apply_rec(x, rec)
}
\arguments{
\item{x}{A dataset to be preprocessed.}

\item{rec}{A \code{recipe} object specifying the preprocessing steps.}
}
\value{
A preprocessed dataset based on the provided recipe.
}
\description{
Preprocesses a dataset using a specified recipe.
}
\examples{
\dontrun{
  # Assuming 'rec' is a predefined recipe object and 'data' is available
  preprocessed_data <- apply_rec(data, rec)
}
}
