% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{create_plot}
\alias{create_plot}
\title{Create a Plot from Model Results}
\usage{
create_plot(x, y)
}
\arguments{
\item{x}{A machine learning model object.}

\item{y}{An object containing data splits.}
}
\value{
A ggplot object visualizing the validation and training errors.
}
\description{
This function takes a model and data splits, computes validation and training errors,
and then generates a plot to compare these errors.
}
\examples{
\dontrun{
  # Assuming lm_model and data_splits are available
  create_plot(lm_model, data_splits)
}

}
