% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analysis_augment.R
\name{extrac_prep}
\alias{extrac_prep}
\title{Extract and Prepare Preprocessor from Model}
\usage{
extrac_prep(model, x)
}
\arguments{
\item{model}{A model object, typically a workflow or recipe.}

\item{x}{A dataset to be preprocessed.}
}
\value{
A preprocessed dataset, where the preprocessing steps defined in the model are applied to \code{x}.
}
\description{
Extracts the preprocessor from a model and applies it to the provided data.
}
\examples{
\dontrun{
  # Assuming 'model' is a predefined model object and 'data' is available
  preprocessed_data <- extrac_prep(model, data)
}
}
