% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_training_error.R,
%   R/get_validation_augment.R
\name{extract_fit}
\alias{extract_fit}
\title{Extract Fit Parameters from a Resampled Model}
\usage{
extract_fit(x)

extract_fit(x)
}
\arguments{
\item{x}{A resample result object.}
}
\value{
Returns the extracted fit parameters from the provided model object.

A list of extracted fitted model objects.
}
\description{
This function simplifies the process of extracting fit parameters from a model object
created using the \code{parsnip} package. It is particularly useful in scenarios where
model objects are embedded within resampled results.

Extracts the fitted model object from each resample.
}
\examples{
\dontrun{
data(iris)
model <- linear_reg() |> set_engine("lm")
resamples <- vfold_cv(iris)
fit_results <- model |> fit_resamples(resamples)
extracted_params <- extract_fit(fit_results)
}

\dontrun{
  # Assuming 'resample_result' is predefined
  fitted_models <- extract_fit(resample_result)
}
}
\seealso{
\link[tune]{extract_fit_parsnip}
}
