% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_validation_aumented.R
\name{get_validation_augment}
\alias{get_validation_augment}
\title{Augment Analysis Data from Resampling Results}
\usage{
get_validation_augment(model, splits)
}
\arguments{
\item{model}{A machine learning model object, compatible with \code{tune::fit_resamples}.}

\item{splits}{Resampling splits object, usually created using the \code{rsample} package.}
}
\value{
A tibble containing augmented data from the analysis portion of the resampling splits.
}
\description{
This function fits a model to resampled datasets, preprocesses the analysis data, and then augments it.
}
\examples{
\dontrun{
  # Assuming 'model' is a predefined model object and 'splits' is defined
  analysis_data <- get_validation_augment(model, splits)
}
}
