% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_validation_error.R
\name{get_validation_error}
\alias{get_validation_error}
\title{Calculate Validation Error for Machine Learning Models}
\usage{
get_validation_error(model, splits)
}
\arguments{
\item{model}{A machine learning model object, expected to be compatible with the \code{fit_resamples} method from the \code{tune} package.}

\item{splits}{An object containing data splits, typically generated by functions from the
\code{rsample} package, used for resampling.}
}
\value{
A data frame containing the R-squared metric for each resample along with an identification of the error type as "Validation".
}
\description{
Computes the validation error for a given machine learning model using resampled datasets.
This function fits the model to each split of the data and then collects and filters the performance metrics, specifically focusing on R-squared (rsq) for validation.
}
\examples{
\dontrun{
  library(dplyr)
  library(tidymodels) # assuming tidymodels includes necessary packages
  data(iris)
  model <- linear_reg() \%>\% set_engine("lm")
  splits <- vfold_cv(iris, v = 5)
  validation_error <- get_validation_error(model, splits)
}

}
\seealso{
\link[tune]{fit_resamples}, \link[tune]{collect_metrics}
}
