% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{plot_theme}
\alias{plot_theme}
\title{Create a Customized ggplot2 Theme}
\usage{
plot_theme()
}
\value{
A \code{ggplot2::theme} object that can be added to ggplot2 plots.
}
\description{
This function generates a customized theme for ggplot2 plots.
It modifies elements such as legend position, axis text, and line colors.
}
\examples{
\dontrun{
  ggplot(data, aes(x, y)) +
  geom_line() +
  plot_theme()
}
}
