% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_training_error.R
\name{extract_fit}
\alias{extract_fit}
\title{Extract Fit Parameters from a Resampled Model}
\usage{
extract_fit(x)
}
\arguments{
\item{x}{A model object, typically resulting from a resampling process
(e.g., using \code{fit_resamples} from the \code{rsample} package).}
}
\value{
Returns the extracted fit parameters from the provided model object.
}
\description{
This function simplifies the process of extracting fit parameters from a model object
created using the \code{parsnip} package. It is particularly useful in scenarios where
model objects are embedded within resampled results.
}
\examples{
\dontrun{
data(iris)
model <- linear_reg() |> set_engine("lm")
resamples <- vfold_cv(iris)
fit_results <- model |> fit_resamples(resamples)
extracted_params <- extract_fit(fit_results)
}

}
\seealso{
\link[tune]{extract_fit_parsnip}
}
