% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_validation_augment.R
\name{get_validation_augment}
\alias{get_validation_augment}
\title{Extract Validation Data and Augmented Predictions}
\usage{
get_validation_augment(model, splits)
}
\arguments{
\item{model}{A machine learning model object, compatible with \code{tune::fit_resamples}.}

\item{splits}{Resampling splits object, usually created using the \code{rsample} package.}
}
\value{
A tibble containing augmented prediction data from the validation set.
}
\description{
Performs resampling on a model and extracts validation data and augmented predictions.
}
\examples{
\dontrun{
  # Assuming 'model' and 'splits' are predefined
  validation_data <- get_validation_augment(model, splits)
}
}
